pyproject.toml
[tool.poetry]
name = "idict"
version = "4.220518.12"
description = "Lazy dict with predictable deterministic universally unique identifiers"
authors = ["davips <dpsabc@gmail.com>"]
license = "GPL"
readme = 'README.md'
packages = [
    { include = "idict", from = "src" }
]

# For more dependencies (sqlalchemy): poetry install -E full
[tool.poetry.extras]
full = ["pandas", "sqlalchemy", "scikit-learn"]

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
orjson = "^3.5.0"
lz4 = "^4.0.0"
ldict = "^3.220128.4"
garoupa = "^2.220128.1"
dill = "^0.3.5"
pip = "^21.3.1"
pandas = { version = "1.3.4", optional = true }
SQLAlchemy = { version = "^1.4.27", optional = true }
scikit-learn = {version = "1.0.1", optional = true}
bson = "^0.5.10"
liac-arff = "^2.5.0"
temporenc = "^0.1.0"

[tool.poetry.dev-dependencies]
autoreadme = "^0.2102.20"
pytest = "^6.2.5"
black = "^21.7b0"
pdoc3 = "^0.10.0"
pypiserver = "^1.4.2"
passlib = "^1.7.4"
pylint = "^2.7.4"
ipython = "^7.26.0"
pandas = "1.3.4"
SQLAlchemy = "^1.4.26"
pytest-cov = "^3.0.0"
testfixtures = "^6.18.3"
scikit-learn = "1.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

#[[tool.poetry.source]]
#name = "dvpypi"
#url = "http://localhost:8080/simple"
#secondary = true
